backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
pandas==2.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.4
